api/requirements.txt
fastapi
uvicorn[standard]
websockets
pydantic
python-multipart
python-dotenv
openai[realtime]
langchain
langchain-openai
langgraph
langsmith
numpy
sounddevice
httpx
